Add basic class diagram to project
The class diagram is a really high level look on the structure of the
project as it should be in the future. This class diagram explains the
basic building blocks of the module as they should be in the wanted
implementation.

This might change in the future, but as for the start of the
development, those are all the building blocks wanted needed for the
project.