Add word_couple module to POC
Create the needed directory structure to hold the code of the POC for
the project.

Add a very basic python object of WordCouple, this object holds a couple
of words, one in each language that will be learned. Currently, this
object has only very basic implementation, it will be extended in the
future, when more functionality will be needed.

This commit was done as part of issue #5.