Add dictionary module to project's POC
Add a basic module of a dictionary to the POC of the project. This
module supports a very basic set of operations on the words it
contains. This module will be updated in the future, getting the
extended functionality when it is needed in the project.

This commit was done as part of issue #5