Add the menu module to the POC
The module of the menu is the module responsible on running all the
command and having the interface with the user.

This implementation of the menu have a very basic interface with the
user: it gets a command, separate it to its different parts (by spaces
and double quotes) and then ran the command in case it has it in its
command.

The implementation of the menu doesn't know which commands it has. It
was done this way to allow the menu be as expandable as possible. The
commands themselves will be added in future commits, all of them will
have the needed data for the menu (command name and command function).

The menu uses input function. This was done on purpose, because the
project is written in python3 and it should not run under
python2. It is UNSAFE to run this code under python2, and should not be
done.

This was done as part of issue #5.