Add the abstract action to the project
Add a module of abstract action to the project. Every command that will
be part of the menu in the project will inherit from this base action,
implementing at least the abstract function of it.

This action has a very basic interface. The module has some variable
that saves the data needed for its help message (COMMAND_NAME, USAGE,
HELP_TEXT, ARGUMENTS), and the command implements a function that will
run once the command should run in the menu.

In this commit, __init__.py files were also added to the new directory
and the original directory. This was done because now the project will
have to import functions from other directories, and they must have this
file in order to be available for import.

This commit was done as issue #5.