Add generic menu actions to the project
The menu have some generic actions that it should have in every project
that it will be part of. Those commands consists of at least the help
command and the exit command.

The help command will print the help on the different commands available
in the menu, or a help specific for a single command.
The exit command will exit the menu (and thus probably ending the
program).

Those commands are generic commands, they are not specific for this
project and in case some day the menu will be exported to some other
project, they can be copied too.

This commit was done as part of issue #5.