Add some specific actions to the menu
This commits add some specific commands to the current project that will
be part of the menu. Those are the most basic commands that there are in
the project, and they will allow the user to use the menu for the most
basic functionality of getting a dictionary, editing it and printing it.
Many more commands will be added in the future.

The current available commands are:
* add_dictionary - Add a new dictionary to the menu.
* add_new_word - Add a new word to the dictionary.
* load_dictionary - Load a dictionary from a file it has been written
                    to.
* print_dictionary - Print the current dictionary.
* save_dictionary - Save the current dictionary to a file on the disk.

Currently, the menu supports only one dictionary at a time. To change
this, all the commands will need to change, to get a single dictionary
from a list of dictionaries instead of getting the only dictionary.

The serialization and deserialize function uses the pickle module to
pack and unpack the dictionary. This module is a security risk in
python. It can run malicious code given certain files. Because this
project is mainly a POC for self use, for now it is good enough, but you
must trust the source of your dictionary files (which you will probably
generate yourself). Maybe an alternative to pickle will be used in the
future.

This commit was done as part of issue #5