Add a basic main for running the menu
Until this commit, all the parts of the menu was created, but they were
separated and didn't interact with each other.

This commit adds the main module to the project. This module is
responsible on creating the menu, adding all the different available
commands to the menu and then running the menu.

After this commit, it is possible to run a basic version of the
project. This version will support only doing very basic actions on
a dictionary, but they work.

Currently, the main module needs to know all the different available
commands in order to add them to the menu. Maybe some day it will be
changed to a better, nicer way of doing it.

This commit was done as part of issue #5.