Make the __str__ method of Dictionary more generic
In the previous implementation of the __str__ method in the Dictionary
class it got got the needed values of the WordCouple by itself and
added it to the string representation of itself.

This commit makes the __str__ method more generic by changing this, and
getting the string representation of the WordCouple object from the
word couple as string, and not as different fields. It makes changes in
the string representation of the WordCouple object easier.

This commit was done as part of issue #5