Add commands to get translations using dictionary
Add two commands to get the translations of words back and forth between
the native and the learned languages.
The commands are:
* translate_to_native_dictionary - Translate a word from the learned
                                   langauge to the native langague.
* translate_to_learned_dictionary - Translate a word from the native
                                    langauge to the learned language.

Those commands are a bit awkward (they are long and hard to
write). Those names were choosen for future compatibility. In the
future, more commands of translate will be added, using different
methods (for example, command line tools), so the shorter names should
remain free for now.

This commit was done as part of issue #5.