Add more functionality to Dictionary object
In the first implementation of the Dictionary class, it has a very basic
limited functionality. This commits add some more functionality to this
object, to let the user do more actions on the dictionary.

This functionality adds two basic functions to get words from the
dictionary. The one is can be used to get a single word from the
dictionary, and the other one is used to get all the words in the
dictionary from a given level.
Those functions will be used in the future.

This commit was done as part of issue #5