Update WordCouple class for better testing
This commit changes the implementation of the WordCouple class, to make
it easier to use.

First it adds a functionality of reset_knowledge. This is a function
that will allow the user to reset the knowledge level of a given word,
in case he wants to learn it again.

The other changes, both in the answered_right function and in
answered_wrong function are meant to change the levels more easily. In
the previous versions, the changes of levels happens when the user has
reached the maximum number of right answers, or has gone below the
zero. This functionality changed, and now the level will decrease only
if the user has reached the maximum number of wrong answers, and not
just zero. This is done to allow the user to have few mistakes before
decreasing its level.

The other change was to handle the case in which the user have a word in
the max known level that will never move down. The previous bug was that
the word could get more and more points, without any real limit. Now,
the words in the max known level will not increase any more.

This commit was done as part of issue #5