Add abstract implementation of a question
A question will be used to test the user on various words. There would
be different questions to the different levels of the words.

This commits add the abstract implementation of a question, with all the
abstract and implemented function of it.

This commit was done as part of issue #5