Add multiple choice and translate questions
Add two types of questions to the project.

The first type of questions are multiple choice questions. In those kind
of questions, the user will have a question with four possible
answers. He will need to chose the right answer for the question.

The second kind of questions are translate questions. In those kind of
questions, the user will have a word in one language, and he will need
to translate it to the other language.

In both implementation of the questions, there are also specific
questions implementations. The specific questions are the classes that
will have the different questions for translating between the different
languages. At all, there are 4 questions:
1. Multiple choice native to learned
2. Multiple choice learned to native
3. Translate questions native to learned
4. Translate questions learned to native

This commit was done as part of issue #5