Add tests to the project
Add a class of test to the module. A test will be assembled from a list
of questions, separated by the different types of the words. When
running a test, the user will be prompt with 25 questions, one after the
other. He will answer all those questions, by doing so updating the
known level of the various words of the dictionary, and at the end he
will get the correctness statistic of his answers.

Aside from adding the Test class, this commits adds the needed code to
run the test as part of the menu. The test have a simple run action, it
will just run the test. This action was added to the menu (through the
change in the main module).

This commit was done as part of issue #5