Fix a bug in the WordCouple decreasing mechanism
The object of WordCouple can change its level by decreasing it, or
increasing it. When the user answers right or wrong a given number of
questions, the object changes its level to a higher or lower known level
according.
When the user answers wrong enough times, the known level of the word
should decrease.

However, there was a bug in the code. When the word had reached the
given number of wrong answers, its known level was re-setted instead of
decreased.
The bug happened because there was a reset of the wrong variable in the
class. The variable that should have been set to zero is the variable of
the number of right/wrong answers in the current level, but the one that
was actually set to zero was the known level of the word.

The fix is a very easy one, it simply changes the variable that is set
to zero from the word_known_level to the stage in the learned level
(_current_stage_in_learned_level).

This commit was done as part of issue #5.