Change help action to use interactive menu
Change the help action to use an interactive menu in order to show the
help on the different topics instead of getting them from the menu
itself. This will make the help command more user friendly, and allow
less typing to get the wanted help on the wanted topics.

Now, when the user will run the help command, it will get a menu with
all the available commands in the menu, and will be able to run the help
on any command from there.

Now, the only argument that the help function is using directly from the
menu is the context of the menu. Any other argument that the function
needs it gets from the user.

This commit was done as part of series of commits that will change all
the interface of the program to be more user friendly. This will be done
using two python libraries, consolemenu and qprompt.

This commit was done as part of issue #5