Change AddDictionaryAction to not get parameters
Until now, the directory action got all the parameters in its
command. Now, all the commands are being changed to get the parameters
in itself, in order to change the menu to use the menu from consulemenu.

The dictionary now uses the qprompt library to get the input
interactively from the user.

This commit was done as part of series of commits that will change all
the interface of the program to be more user friendly. This will be done
using two python libraries, consolemenu and qprompt.

This commit was done as part of issue #5