Change AddNewWordAction to not get parameters
Change the action of GetNewWord to get the parameters directly from the
user, instead of getting it from the menu. Now, the action will ask the
parameters from the user, getting it only after the command was run.

The command now uses the qprompt library to get the input
interactively from the user.

This commit was done as part of series of commits that will change all
the interface of the program to be more user friendly. This will be done
using two python libraries, consolemenu and qprompt.

This commit was done as part of issue #5