Change PrintDictionaryAction to use consulemenu
Until now, the PrintDictionaryAction simply printed the dictionary to
the screen. This commit changes it, to make the dictionary use a console
menu to print it to the screen.

This commit was done as part of series of commits that will change all
the interface of the program to be more user friendly. This will be done
using two python libraries, consolemenu and qprompt.

This commit was done as part of issue #5