Change RunTest to use qprompt for input
Change the input method of the RunTest action and the modules it uses to
use the qprompt library to get the input, instead of getting the input
using the `input` function. This allows for more unified input across
the different modules in the project.

This commit was done as part of series of commits that will change all
the interface of the program to be more user friendly. This will be done
using two python libraries, consolemenu and qprompt.

This commit was done as part of issue #5