Refactor actions to stop raising exceptions
Until now, when the actions encounter some kind of problem or an unknown
state, they raise an exception. Later, the menu catch this exception,
print its value and discarded it.
Now, the menu is about to change to the be implemented using
consolemenu, and this kind of menu does not catch the exceptions. To
handle this case, the implementation changed to stop raising exceptions,
instead printing the wanted error.

This commit was done as part of series of commits that will change all
the interface of the program to be more user friendly. This will be done
using two python libraries, consolemenu and qprompt.

This commit was done as part of issue #5