Change menu implementation to use consolemenu
Change the implementation of the menu to use the library of consolemenu
in order to get have the menu, instead of implementing the menu by
itself.

It is done because this menu library is better than the simple
implementation of the menu that was done before.

As for this commit, the code is fully working with the new menu, but
there is still a lot of unused interface that was written for the
previous menu. The next commits will remove all this code.

This commit was done as part of series of commits that will change all
the interface of the program to be more user friendly. This will be done
using two python libraries, consolemenu and qprompt.

This commit was done as part of issue #5