Remove unused code from all actions
The actions were originally built for the menu implementation that were
written by me in this project in the past. Now, the menu has changed,
and the project is using an open source library for the menu
implementation (consolemenu).

The previous commit has changed the menu to start using the interface of
the consolemenu. This commit changes the interface of all the actions to
be as expected by the library.

The actual changes are as followed:

* The functions are no longer getting parameters in args and xargs. The
  menu doesn't send the command line parameters to the functions
  (because it doesn't get them), so the functions doesn't get them from
  there anymore. All the actions are getting those parameters by asking
  the user to insert them (as was changed in previous commits).
* Exit action was remove - The menu has a built in exit function, so the
  action is no longer needed. This module has been completely removed
  from the project.
* Functions are no longer returning value - In the previous
  implementation, the exit was implemented by using a return value from
  functions, when a given functions has returned False the menu
  stopped. Now, the menu is using the open library that have its own
  interface for closing, so it is no longer needed. All the actions are
  now not returning values from them.
* ARGUMENTS and USAGE documentation were removed - as explained earlier,
  the functions are not getting any arguments any more, so the USAGE and
  arguments are redundant now. They were removed from all the actions
  and from the base implementation that had them in the message format.
* Command name changed - Until now, the command name was in a single
  word (in snake_case). This was done to make the menu easier to use and
  separate between the commands and the arguments. Now, the menu handle
  it by itself, so it is not longer needed.

This commit was done as part of series of commits that will change all
the interface of the program to be more user friendly. This will be done
using two python libraries, consolemenu and qprompt.

This commit was done as part of issue #5