Reorder the commands in the menu
Change the order of the commands in the menu. At first, the commands
order was random, because it didn't affect anything. However, now, when
changing the menu to the consulemenu, the order of the commands are
important, because it is printed that way.

Now, the commands are more ordered, and should be easier to use.

This commit was done as part of series of commits that will change all
the interface of the program to be more user friendly. This will be done
using two python libraries, consolemenu and qprompt.

This commit was done as part of issue #5