Add default value for dictionary saved name
This commits add a new functionality to the project. Until now, when
saving the dictionary to the disk, the user always had to type the name
of the dictionary by hand.

This commit saves the name of the dictionary in case it was loaded by
the program, and it set this name as the default value for the name of
the dictionary in case it should be saved. This should save the user the
typing of the name of the dictionary most of the time.