Move the design to the C directory
Since the project has been change, and from now on it won't be only
C implementation of this project, every implementation should be in its
own sub-directory in the project.

The design directory was intended to be specific for the C language, and
won't fit to the other languages implementation, so it moved to the
C directory.

This commit was done as part of issue #8