Move all python files from POC to Python
Since the project has been change, and from now on it won't be only
C implementation of this project, every implementation should be in its
own sub-directory in the project.

The python project started as a proof-of-concept project, under the
utils program, but now that the project has many implementations in it,
it deserves a directory in the root of the project.

This commit was done as part of issue #8