Update README file with full project description
Update the README file with a full description of the project. Now, the
README can explain to any new user what this project is all about and
how he should use it.

The README is now full for the current implementation of the project,
from now on, it should be update together with updates from the code.

This commit was done as part of issue #7