Add abort option to add_new_word action
The action of add_new_word is an action responsible on adding new words
to the dictionary of words. Until now, when this command run, the user
had no way to abort the action.

It means that after running this command, the user had to add a new word
to the dictionary, or exit the program altogether. This was an issue
when the user changed his mind about adding a new word, or had an error
while inserting one of the words (for example, in case of a typo in the
native word).

This commit fixes this issue, by adding an option to abort this action
even after choosing it. This commit adds let the user abort the
add_new_word action. It is done by inserting -1 in one of the fields in
the action. This was chosen this way because -1 can never be any valid
value in any field.

I tried to add option to exit by sending keyboard interrupt as well
(<c-c> on most systems). However, it seems that the menu catches this
interrupt and crashes, so it won't work.

This commit fixes issue #12