Move add_new_word logic to static function
Until now, the run function of the command add_new_word had done all the
logic in itself. It got the dictionary from the context, ask the user to
insert the input and then inserted the input into the dictionary.

This commit changes this. Now, the run only extracts the dictionary from
the context of the menu, and call an internal static function that does
all the other logic (getting the words from the user and adding them
into the dictionary).

This was as part of issue #11. This issue is going to add a new command,
that will allow the user to add multiple words to the dictionary one
after the other. To avoid code duplication, the new command will just
call the function that already adds a new word. Since no such function
existed, it was created in this commit.