Create BulkAddNewWordAction action
Create a new action in the project. This action is used to add a bulk of
words, instead of just adding a single word every time, when the user
wants to add a lot of words one after the other.

This command will be very useful when adding a lot of words at the same
time, it will save the user a lot of timing and running the
AddNewWordAction over and over again.

This action works by using the functionality exported from
AddNewWordAction. The new command simply calls the add_new_word function
over and over again, as long as the user wants to add more words to the
dictionary. Once the user will not want to add the current word (by
setting one of the words values to be -1).
This might make the user sometimes exit the command before he wants to
(for example, if he had typo in the first word, and it wants to abort
only this word), however, it seems like the easiest interface to use, so
it was chosen.

I thought about creating this command as an inheritance of some new
abstract command that is going to run for all the actions that wants to
works in bulks. However, as the code is written now, it seems that
adding such an interface is really redundant - all it will save is the
loop call to while(<inner_action>). It is the case because of the
interface of the commands (which can not get different parameters from
each other in their initialization) and because the action of the bulk
add is very very simple. In case one of those will change (the interface
of creating commands or the number of same code that actually runs in
the bulk command) it might be wise to change this implementation to be
as described above.

This commit fixes issue #11