Add option to ask for number of question in test
Until now, when the user wanted to run a test on his vocabulary (by
using the command of RunTestAction) he run the command, and got a test
that had 5 question in each level of words.

This commit adds an option to control the number of question per level
the use will get. Now, after running the RunTestAction action, the user
is prompt with a question about the number of question in a level he
wants. The default answer to this question stays 5 (the previous number
of questions per level as was before), but, now the user can set this
number to any value he wants, and this value will be the number of
questions per level.

There was already support for this feature in the class of Test,
however, it was not used in the previous commits (ever since changing
the menu to use consulemenu). Now, this feature is used again.

This commit fixes issue #17