Add basic dictionary app
This is the first commit in the Kotlin development of the project. This
commit adds a very simple application that only contains a very basic
dictionary.

The current application saves a very basic dictionary. This dictionary
has an action that allows the user to add new words to it, and has an
action that allows the user to add new words to the dictionary, and
print the dictionary.

The current implementation saves the dictionary between the runs of the
app, which means that the dictionary is persistent. The dictionary has
also an option to clear itself completely. This is only a basic
functionality, and should be updated in the future to have options to
delete single words from the dictionary.