Create working version of the code as application
This commit creates a completely working version of the python code as
an application that can run on phones. It completes the work that
started in the previous commits, that converted tho code to Kotlin for
Android.

This commit adds all the actions relevant to test and questions in the
application. A lot of the code looks ugly, many places still have TODO
in them, with fixes that should be done. Most of them should be fixed in
the next commits of the project.
A lot of the Kotlin code doesn't look good, and doesn't follow the best
practices of the language. I am still working on getting familiar with
the language, so some code is written wrong for this language.

This commit was done as part of issue #19