Add picture files to the app
To upload the app to the app-store I had to add logo and screenshots
from the app. This commit adds those to the repository.

The current logo is a very bad one (five minutes of work in Paint), it
should be update by a more skilled designer in the future.