Add implementation for safe-ask action
Add a new module with a single method that implements an action that can
get a string safely from the user.

This method will try to run the given ask method until it will get
a value back from the user. In case the method will raise the exception
of UnicodeDecodeError (which happens when the input is invalid), the
command will ignore the input from the user and try to get a new one
from it.

This commit was done as part of issue #32