Change all ask_str to use safe_ask instead
This commit moves over all the places in the code with a call to ask_str
and changes them to be using safe_ask with the ask_str there
instead. This makes the calling to this function safe, and make sure
that the program won't crush when it is used.

This was done as part of issue #32