Change all ask_int to use safe_ask instead
This commit moves over all the places in the code with a call to ask_int
and changes them to be using safe_ask with the ask_int there
instead. This makes the calling to this function safe, and make sure
that the program won't crush when it is used.

This was done as part of issue #32