Fix the bug to get right but unexpected for learned
There was a bug that prevent the code from returning that the given
answer is correct but not the expected one in case the answer is in the
translated language.

The bug was that the wrong function was called in this case. In stead of
calling the function that returns all the words in the translated
language (as they should be), the function that was called was the
function that returns all the native translation of the given word. This
bug was fixed by calling the right function in the dictionary.

This commit fixes #35