Change native_word variable from public to private
The variable of native_word inside the class of WordCouple was a public
variable up until now. This commit changes it, and make it a private
variable instead. This is done because the implementation of this
variable is going to change in the near future, and it should be
internal to the class and not public.

The commit adds a getter method for this variable, and changes all the
code that uses this variable to use the getter method instead of using
the variable itself.

This commit breaks the internal variables of the class of WordCouple,
which means that dictionaries that were created before this commit won't
load any more after this commit.

This commit was done as part of issue #38