Change native word in WordCouple into a list
Until now, the native word in the word couple was a single
variable. This commit changes it and set this variable as a list instead
of as a single string.

This change will allow future commits to add more than one native word
in a word couple, as the issue wants to solve.

Right now, the commit only change the implementation of the word couple,
but does not add any functionality to it. It means that the user still
won't be able to add more than one word as a native word, because there
are still no interface functions that let him do it.

This commit was done as part of issue #38