Move word implementation to its own class
Until this issue, every word in the word couple was only a regular
string in the program. This issue makes the word more complicated the
this - every word can have various spellings options and they can have
some functionality that should be in them.

Until this commit, there was code duplication between the different
words in the word couple (translated and native). This commit moves this
functionality to its own class, so there won't be any code duplication
between the different words in the word couple.

As a result, there is no reason to leave the words in the words couple
as private, so they turned back to be public variables in this class.

The commit also adds the functionality of list of words to the
translated word. The previous commit added this functionality to the
native word only, but now the words are implemented using a common
class, so both of them has the same functionality.

This commit was done as part of issue #38