Add Word and WordCouple equal checking methods
Add some functions that let other objects in the code better check if
different words are the same as one another. The new functions can check
if a given word is the same as a string or as another word, and if
a given word couple is the same as two string words or the same as
a another word couple.

This commit makes all the places in the code that checks for equality
between words look a lot better.

This commit was done as part of issue #38