Refactor add_word implementation in Dictionary
This commit refactors the function of add_word inside the class of the
dictionary. This function has gotten too big, so this commit changes its
implementation and creates some internal functions that are used in
order to do the add_word function functionality.

This commit was done as part of issue #38.