Add option to extend existing words
Until now, every word has a single value that was set to it in every
language. The previous commits added the functionality that allowed the
words to have more than one spelling option. This commit changes the
interfaces of the classes to let the user use his feature.

The commit adds option to add spelling to words. Until now, the words
has a list of spelling, but there was no way to alter this list. This
commit adds two functions that alter the list of words. The first one
adds the new spelling as a secondary spelling to the word, the second
adds the new spelling as a primary spelling to the word.

The commit add a check function to the WordCouple object. This function
checks if two words are close to one another. This function can be used
to check if two words are actually the same word, with a different
spelling. This function checks the closeness of words by checking that
they have at least one common value in them (the word in one of the
languages is the same). It does not actually moves on the words and
checks the different between them.
The commit adds option to extend a given word in the word couple. This
function adds a new word into a given word couple. By doing so, it
merges all the spellings in both words.

The commit changes the implementation of the add_word function inside
the dictionary. Now, before this functions adds a new word to the
dictionary it checks if a close word exists in the dictionary. In case
a close word exists in the dictionary, it prompt the user with question
about whether those two words are the same. In case the user says that
these words are the same, the dictionary adds the new word as part of
the previous word, instead of as a new word. This implementation still
needs a little bit of refactor, and a todo was left in the code.

This commit was done as part of issue #38