Fix linter errors in project
This commit moves over all the files in the project, running linters on
all of them. It fixes all the wanted errors that the linters warn
about. There are more errors that the linters report on, but they are
ignored intentionally.

The linters that run in this commit are:
 * pylint
 * bandit
 * flake8
 * frosted
 * mypy
 * prospector
 * py3kwarn
 * pycodestyle
 * pydocstyle
 * pyflakes
 * pylama
 * python

This commit was done as part of issue #38