Add load words file action to menu
Add a new action to the menu. This action reads a yaml file with
a list of dictionaries and load them into the current dictionary.

This action will allow the user to automate adding of new words to the
dictionary, making it a lot easier to add new words to it.

Fixes #41.