Add command to remove words from dictionary
Add a new command to the project, that removes a given word from the
dictionary. This command can be used in order to remove words that were
added to the dictionary by mistake, or that contains errors.

The current command removes the words according to the index of the word
in the dictionary. This is done this way because this is the easiest way
to remove the words, it can be changed in the future to support better
removing.

This commit fixes #13